it now reads/writes to same file
Now the code will read and write to the same file. It will write the TA
to a different tab and re-save the file in the end, provided you change
the area it grabs from the time series tab (sheet 1) and change the tab
it writes things back with on line 115.

Next, I'll try to just have it read all the data and output it all to
the same tab. Baby steps, though.